requirements.txt
fastapi
sqlmodel
passlib
pydantic-settings
python-jose[cryptography]
python-multipart
pymysql
uvicorn[standard]
pydantic[email]